moving queries filed to client root
